make sure FileCheck is copied in the LLVM output directory
The tool, which is needed by parts of our test suite, is built as part
of LLVM but is *not* copied to the directory containing the output LLVM
binaries. This adds a flag to ensure the binary is copied.

This shouldn't add any extra built time, as the flag just installs extra
binaries that were already compiled.