asm/arm: allow r9 when usable, make diagnostics more specific
As described in #85247, there are a number of cases where inline ARM assembly
in Rust programs is prevented from using certain registers. Sometimes these
registers are only reserved in certain situations; however, they are all
unconditionally restricted by the compiler. Many of rustc's diagnostics
describe the registers as being "used internally by LLVM", which obfuscates
the meaning of the registers and is ultimately unhelpful to low-level
programmers.

At the moment, it does not seem possible to ensure that the frame pointer or
base pointer will be elided, or to understand the conditions under which this
may occur. Some of this is due to the addition of a codegen abstraction layer,
where these behaviors may vary between backends and may not be configurable.
As such, this information is not available to Rust's assembly parser.

However, register r9 has a condition that is rather easily checked: the
relocation model. r9 is used as the "static base" (SB) for "read-write
position independence" (RWPI) as a base from which to address writable data,
as defined in the Procedure Call Standard for the Arm Architecture[1]. The
register is always available for use if RWPI is not used[2].

Rust allows RWPI mode to be selected on a per-target basis, as well as making
it available as a codegen option in rustc. Thus, we can predicate usage of r9
on this setting with confidence.

Allow r9 to be used in `asm!` blocks when the target does not use RWPI, and
define what r6 and r9 are reserved for in their respective diagnostics. The
frame pointer (r7/r11) and base pointer (r6) remain unconditionally restricted
for now.

[1]: https://developer.arm.com/documentation/ihi0042/j/
[2]: https://developer.arm.com/documentation/dui0056/d/using-the-procedure-call-standard/read-write-position-independence/register-usage-with-rwpi