Disable projection sub-obligation optimization in intercrate mode
The meaning of an 'evaluation' in intercrate mode is more complicated
than with an normal `SelectionContext`. To avoid potential issues,
we now always preserve all projection sub-obligations when in
intercrate mode. This avoids needing to answer whether or not
`EvaluatedToOk` always means the same thing in intercrate mode
as it does normally.