Add `impl From<Ordering> for i32`
This commit adds a first-class `From` conversion of `Ordering` into
`i32`.

`Ordering` is defined to be compatible with the return types of the
various `...cmp...` routines in the C standard library, but the only way
to convert an `Ordering` to an `i32` is with a non-obvious `as` cast.

This commit adds a proper API for doing this and hides a "scary" `as`
cast inside the `core` library.