Ship an llvm-coverage-tools component with just llvm-cov and llvm-profdata
People use the llvm-tools-preview component for coverage tools, and
documentation refers to those tools, but we don't necessarily want to
stably commit to shipping all of those tools. Split out a separate
component for llvm-cov and llvm-profdata. Ship them in the bin directory
to make them easily runnable, to reduce the need for wrappers to find
and run them.