Add a visibility info tag, and remove pub
For most documentation, the `pub` in front of each method is redundant:
you know the method is public because you are reading the documentation.

However, when running with `--document-private`, it's important to see
which items are public and which are private. This adds an item-info tag
for private items showing their visibility, and removes the `pub` from
methods and fields.

This doesn't currently touch the summary code block at the top of the
page, which will still have `pub` or `pub(crate)` or whatever is
appropriate.