Preliminary support for verbatim linking of crates
Due to the fact that if we were not using verbatim linking, linking
against a versioned crate will only work if a unversioned symlink to
the versioned one is part of the searchpath.
This is "preliminary" as right now it only uses the whole filename
for GNU-style linkers. It should however, be expanded to any linker
that supports verbatim linking.

Also added suggestions from @wesleywiser.